figure out timezones.  Caused a lot of readme writing.
I hate speculating this much without Writing Some Dang Code, but this
is thorny as heck and I think hacking something out and trying to nudge
it towards correct will actually be harder than doing some deep
principle research first and just making some decisions.

SO here's a gagload of notes on that.