i'm not sure if this is an idiomatic way to do things, but I got error handling to work.
I'm honestly pretty surprised if there's not a better idiom in typescript for saying
"result or error", but I guess I can kinda see how life ended up like this,
given that things are still just whamming structural typing on data that is oblivious.

kind of a shame there's no syntactic sugar for making a sum type with discriminators, though.