halfway through doing a new take on the types and validation approach.  Moving away from zod.
I'm not sure if I'm holding it wrong (and I very probably am),
but I've kinda found zod to be less ergonomic than I was hoping.

(And Zod is also very nontrivially expensive to import at all:
it appears to be increasing my bundle size by about 2 whole megabytes,
which is more than 50% of the whole thing.  I suspect this results
in nontrivial plugin loading time increases, and yes, I very VERY
much do care about that.)

I have a couple of wants:

- it seems like I typically do want the parsed values
  (and yes, I concede, I might as well hold on to Luxon most of the time);

- it seems like I _also_ constantly want the raws
  (sure, I can format-strings my way back to them, but it's *really* *constantly*,
  because things pretty much get stringified on their way to the DOM, from the DOM,
  to the frontmatter, or from the frontmatter.  The parsed lifespan is *short*);

- I want to aggregate any errors;

- and it's not been super helpful to have a value that I can't create in
  the case of an error somewhere within it.  That's just been megapainful.
  I'd rather have something I can accumulate WIP data in, and validate later.
  (Yes, I know about 'Partial<Foo>' now, and that's very nice, but...
  I think going further and retaining errors will be handy too.)

- and I want some place to stash methods.
  (Or at least

And of all the things about Zod that make me itch, the fact you can't
attach methods to a type that's inferred from Zod... well, again, maybe
I'm "holding it wrong", but... that seems... hmm.  Unfit.

So I broke down and implemented a new class for HCEvent.

And then I found myself hitting the raw-vs-parsed and
plz-accumulate-error things so much I created a Control type
to wrap that, generically.

And then by total coincidence I found that while trying to make
a little bit of code reuse in the event editing modal... having
a handle that wraps the raw value makes it *much* easier to extract
certain features, _without_ having to do metaprogramming to poke
at a field of an object.  (Which is what I _was_ originally planning
to do.  But... well, why?  When _most_ of our fields got this
Control wrapper type already, might as well keep rolling with it.)

I'm not sure all this is perfect.

It hasn't used "branded" types at all, and I wonder if I could be
even fancier if it did.

And I'm especially a little worried I prematurely optimized the
Settings creation DRY code, heh.  (But that's easy to pull back from.)

I think this kinda works, though.  I think this is going well.