fully yeet zod, and succeefully fully parse our own data.  Including adding an Option helper type.
I was wrong about the bundle size; it only dropped from 4M to 3.6M.
It seems that esbuild is vendoring all of luxon, all of dataview...
a lot of things it that doesn't need to and really definitely should not,
considering I was trying to get them from dataview at runtime.
Not yet familiar with how to fix that.