ask fullcal not to do event overlaps at all.
If it was a little more aggressive at shrinking them, it might be fine.
But the actual behavior appears to be:

- If there are two events, the second one gets 50% width.
  The first one stays *full width*.

- If there are three or more events, NOW all of them start getting
  compressed... to various degrees.

If it was even that the first event gets at least a 20% width reduction
in all cases of overlaps, I'd probably be fine with it.  But zero... no.
Looks super confusing to me.

And there doesn't appear to be an overlap weight parameter, so.