tell fullcalendar to go 100% height instead of auto.  Different practical results.
Telling it to do "100%" causes it to take up the container reasonably,
but also causes the scrolling to be internal to the FC elements,
instead of relying on the container's scrolling...

Which has the very important outcome that `scrollToTime` function actually works.

Also, the all-day region stays pinned at the top when you scroll,
which I guess is arguably correct.  (I'm not sure I minded it scrolling away,
either, personally, but.  Eh.

Oddly it made some other CSS around the day labels _stop_ working, but
you know what, I'm not even going to ask about that.