Quickest, dumbest UI for listing categories on an event in the edit modal, before going deeper.  Placeholder.
I'm beginning to see the appeal of frameworks that do data binding.
Doing this the dumb way with very direct wiring is still a viable
quantity of mental gymnastics at this UI complexity level, but I'm
only ever going like one level deep, with a single extra element
to keep updated, which is easy mode.