a little more documentation about what some of our styling is doing.
I'm finding a desire to add custom styling to... basically every modal,
because a considerable number of interactions are going through them,
and having them all be sized the same when there's a series of them
in play is very confusing.

This is kind of unfortunate, given how modals on mobile also need a
whole separate set of styling, because their defaults are so different.
Ah well.