okay, trying some new things for calendar view size controls.  so far, what not to do.
right now the built-in buttons from fullcalendar are doing all the lifting,
but I'd rather declare independence from them.  They're not exactly the
semantics I want (fixed sizes??  no??), and they are in bad need of styling
anyway (they... very much do not fit on the screen in the mobile context,
which is quite oof, and I've just been ignoring it so far).

the "menu" component is... not what I want, i'm afraid.

so it looks like we'll be making something slightly custom.

this ain't it, but i'm starting in the direction.