alright, overdue to start extracting some of the controller logic from the view.
leaving the code in the UI folder because it seems very closely coupled
to the calendar itself.  a decision subject to change, though.

no actual logical changes yet (although i've started to see some places
that probably do need it... such as the exact order of events around
where this newly-extracted function is called).  one step at a time.