add validity check.  knocks out a TODO.
haven't actually seen this one be particularly relevant in practice,
but it seems like it ought to be there, so now it is.

(i'm actually not 100% sure this makes sense to be strict on,
since _probably_ the change can be applied even to invalid data,
and that doesn't make the data any MORE invalid, and I think it'd
be in line with our conventions about sussy data in this project
to actually roll with that.  But: let's try strictness and see if
that ever causes me any encounters that prompt review on this again.)