I'm not sure this will be the place for these style tweaks, but: make keyboard focus MUCH more visible.
I haven't researched if this is my theme or base stuff, and _surely_
this should be parameterized better than this... but starting now,
I really want to SEE where focus is at.

I especially care about this in the modal for category selection,
but just generally across the board, I want horizoncal to be easily
keyboard navigable.

Without this tweak, the way I'm experiencing it is that there's a TINY,
TINY enhancement of the size of a grey(-on-grey!) border around focused
elements.  It's basically invisible unless you see the change in motion,
and even then it's darn hard to see, and invisible again as soon as you
move your eyes.  Poor.

The transition I'm removing here: isn't really visible in the
grey-on-grey situation.  But it becomes _extremely_ visibly jarring
when the more visible highlight color is added.