jankiest possible min poc of expanding view range.
notice that the 'viewType' param to changeView() is *not* the same
strings as are keys in the 'views' map when creating the calendar...
so... `this.calUI.changeView(this.calUI.view.type, ...)` is...
not actually correct.  That's a bit of a head-scratcher.

I could write a case switch for reducing those back into the domain
that is accepted, but I think I ignore this for now, because
at some point we'll just... remove those custom view types, since
this feature makes them irrelevant.

We'll probably want to switch on the view type anyway at that point,
because these buttons should have different behaviors for month-scale
views.

But I think we'll worry about ALL of these things... after I make
the fullcal view range part of the obsidian view's state object, so
any of this sticks, and thus the whole thing becomes more relevant.
And after I finish the rest of the button UI so the old one can be
discarded.  So.  Later.