jankiest possible first implementation of letting categories set events to render with reduced opacity.
I can't see a way to add css style literals, which is probably how I
would've preferred to do this, simply because that would let me easily
proxy a user's exact choice of opacity number through all the way.
But custom classes are possible, so, I make one of those and...whee!

I guess maybe the endgame will be just introduce a view more of those
for 10/20/30/etc steps of ten, and that'll have to do.

This DOES have the desired effect on both timegrid AND list views,
so that's pleasant.

I might want to stack even more effects on for the semantic I'm using
this for, like strikethrough, or mutating the background to have a
gradient (in grid view, at least), etc... but we'll see.

I _don't_ want to make one css class for "cancelled event" and be
prescriptive like that about categories.