Satisfy the rest of eslint.  It's almost all "let"->"const" changes.
Having done all of these, I *really* don't see the point.

Opinions may vary, but to me, calling something a "const" and then
calling "push" on it in the immediately subsequent lines is just
actively confusing.

I'm still going to leave it on for now, for shrug purposes, but...
I'm unlikely to ever enable this eslint behavior on any other project.

I don't think any of the other comments its made are useful either.
It provides one underline of an entirely unused variable... Big whoop.